Rename metrics2 to metrics
